requirements.txt
# Core dependencies
python-dotenv>=1.0.0
httpx>=0.25.0
mcp[cli]>=0.1.0

# LLM
openai>=1.0.0

# Database drivers
mysql-connector-python>=8.0.0
psycopg[binary]>=3.1.0